manifest script
